// api/meta-depot.v1.d.ts
/**
 * This file was auto-generated by openapi-typescript.
 * Do not make direct changes to the file.
 */

export interface paths {
    "/minecraft-versions.json": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        get: {
            parameters: {
                query?: never;
                header?: never;
                path?: never;
                cookie?: never;
            };
            requestBody?: never;
            responses: {
                200: {
                    headers: {
                        [name: string]: unknown;
                    };
                    content: {
                        "application/json": components["schemas"]["MinecraftWithNeoForgeVersionListing"];
                    };
                };
            };
        };
        put?: never;
        post?: never;
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
}
export type webhooks = Record<string, never>;
export interface components {
    schemas: {
        MinecraftWithNeoForgeVersionListing: {
            latest_snapshot?: components["schemas"]["MinecraftWithNeoForgeVersion"];
            versions: components["schemas"]["MinecraftWithNeoForgeVersion"][];
        } & {
            [key: string]: unknown;
        };
        MinecraftWithNeoForgeVersion: {
            version: string;
            /** @default release */
            type: string;
            /** Format: date-time */
            released: string;
            neoforge_version: string;
            /** Format: date-time */
            neoforge_released: string;
        } & {
            [key: string]: unknown;
        };
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}
export type $defs = Record<string, never>;
export type operations = Record<string, never>;